Add iflix experience
If you're part of the iflix team and you're reading this as a result of
an automated check on the CVs of the employees... do not worry. I'm not
looking for a new position :)

We got an email asking for an updated CV (why?!) and I'm updating mine
as requested. On company time, which I think is a giant waste, but the
email seemed very urgent...